feat(splash): Add background image asset for Splash Page
Added `splash_background.png` image asset to `assets/images/` for use in the Splash Page.

This commit introduces the background image asset that will be used to visually enhance the Splash Page.

The `splash_background.png` image will provide a visually appealing background for the Splash Screen, improving the initial user experience when launching the application.