chore(core-config): Update baseURL in ApiConfig to new backend endpoint
Updated the `baseURL` constant in `lib/core/configs/api_configs.dart` to point to the new backend API endpoint: 'https://movieapp-production-3fa6.up.railway.app/'.

This commit updates the application's API configuration by changing the base URL to the new backend endpoint.

This change may be due to:
- Deployment of the backend to a new hosting environment.
- Switching to a different backend environment (e.g., staging, production).
- Migration or update of the backend server infrastructure.

Ensure that the new `baseURL` 'https://movieapp-production-3fa6.up.railway.app/' is the correct and intended endpoint for the application to communicate with the backend API.