feat(home-ui): Implement tappable MovieCardWidget with navigation to MovieWatchPage placeholder
Implemented tappable `MovieCardWidget` in movie carousels on the HomePage, enabling navigation to a placeholder `MovieWatchPage` for movie details and watching.

This commit enhances user interaction and begins implementing the movie watch flow by:

- **Making `MovieCardWidget` tappable with navigation:**
    - Implemented `onTap` callback in `MovieCardWidget`'s `GestureDetector` to enable tap interactions on movie cards.
    - Added `AppNavigator.push` navigation to the `onTap` callback to navigate to the new `MovieWatchPage`, passing the `movieEntity` data as a parameter.

- **Implementing placeholder `MovieWatchPage`:**
    - Created `MovieWatchPage` in `lib/features/watch/presenation/pages/movie_watch_page.dart` as a placeholder page for movie watching functionality.
    - `MovieWatchPage` currently has a basic `Scaffold` structure with an `AppbarWidget` (showing back button) and an empty body.
    - `MovieWatchPage` constructor takes a required `MovieEntity movieEntity` parameter to receive movie data for the details page.

These changes enable user interaction with movie carousels on the Home Page, allowing users to tap on movie cards and navigate to a dedicated `MovieWatchPage` to view movie details and potentially watch the movie in the future.  The `MovieWatchPage` serves as a placeholder for the upcoming implementation of the full movie watching experience.