feat(core, di): Register movie and television movie search use cases
This commit registers the `SearchMovieUseCase` and `SearchTelevisionMovieUseCase` in the service locator (GetIt).

Following the previous refactoring that clarified 'Television' to 'TelevisionMovie' and separated search functionalities, this commit makes the newly created search use cases available for dependency injection.

By registering these use cases as singletons in `service_locator.dart`, they can now be easily injected and used in other parts of the application where search functionality for movies and television movies is required.

This step is essential to enable the implementation of the search feature in the UI and other application layers.